+++ src/main/java/net/geforcemods/securitycraft/blockentities/ReinforcedDispenserBlockEntity.java
@@ -1,0 +1,148 @@
+package net.geforcemods.securitycraft.blockentities;
+
+import java.util.EnumMap;
+import java.util.Map;
+
+import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.api.IModuleInventory;
+import net.geforcemods.securitycraft.api.INameSetter;
+import net.geforcemods.securitycraft.api.IOwnable;
+import net.geforcemods.securitycraft.api.Owner;
+import net.geforcemods.securitycraft.inventory.InsertOnlyInvWrapper;
+import net.geforcemods.securitycraft.misc.ModuleType;
+import net.geforcemods.securitycraft.util.BlockUtils;
+import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.block.state.IBlockState;
+import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.network.NetworkManager;
+import net.minecraft.network.play.server.SPacketUpdateTileEntity;
+import net.minecraft.tileentity.TileEntityDispenser;
+import net.minecraft.util.EnumFacing;
+import net.minecraft.util.NonNullList;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.TextComponentString;
+import net.minecraft.world.World;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.items.CapabilityItemHandler;
+
+public class ReinforcedDispenserBlockEntity extends TileEntityDispenser implements IOwnable, IModuleInventory, INameSetter {
+	private InsertOnlyInvWrapper insertOnlyHandler;
+	private NonNullList<ItemStack> modules = NonNullList.withSize(getMaxNumberOfModules(), ItemStack.EMPTY);
+	private Owner owner = new Owner();
+	private Map<ModuleType, Boolean> moduleStates = new EnumMap<>(ModuleType.class);
+
+	@Override
+	public NBTTagCompound writeToNBT(NBTTagCompound tag) {
+		super.writeToNBT(tag);
+
+		if (owner != null)
+			owner.save(tag, needsValidation());
+
+		writeModuleInventory(tag);
+		writeModuleStates(tag);
+		return tag;
+	}
+
+	@Override
+	public void readFromNBT(NBTTagCompound tag) {
+		super.readFromNBT(tag);
+
+		owner.load(tag);
+		modules = readModuleInventory(tag);
+		moduleStates = readModuleStates(tag);
+	}
+
+	@Override
+	public boolean shouldRefresh(World world, BlockPos pos, IBlockState oldState, IBlockState newState) {
+		return oldState.getBlock() != newState.getBlock() && oldState.getBlock() != SCContent.reinforcedMovingPiston; //prevent this TileEntity from getting removed when the previous block was a Reinforced Moving Piston to make Reinforced Pistons work
+	}
+
+	@Override
+	public boolean enableHack() {
+		return true;
+	}
+
+	@Override
+	public ItemStack getStackInSlot(int slot) {
+		return slot >= 100 ? getModuleInSlot(slot) : super.getStackInSlot(slot);
+	}
+
+	@Override
+	public ITextComponent getDisplayName() {
+		return hasCustomName() ? new TextComponentString(customName) : getDefaultName();
+	}
+
+	@Override
+	public ITextComponent getDefaultName() {
+		return Utils.localize(SCContent.reinforcedDispenser);
+	}
+
+	@Override
+	public <T> T getCapability(Capability<T> capability, EnumFacing facing) {
+		if (capability == CapabilityItemHandler.ITEM_HANDLER_CAPABILITY)
+			return BlockUtils.isAllowedToExtractFromProtectedBlock(facing, this) ? (T) super.getCapability(capability, facing) : (T) getInsertOnlyHandler();
+		else
+			return super.getCapability(capability, facing);
+	}
+
+	private InsertOnlyInvWrapper getInsertOnlyHandler() {
+		if (insertOnlyHandler == null)
+			insertOnlyHandler = new InsertOnlyInvWrapper(this);
+
+		return insertOnlyHandler;
+	}
+
+	@Override
+	public ModuleType[] acceptedModules() {
+		return new ModuleType[] {
+				ModuleType.ALLOWLIST
+		};
+	}
+
+	@Override
+	public NonNullList<ItemStack> getInventory() {
+		return modules;
+	}
+
+	@Override
+	public NBTTagCompound getUpdateTag() {
+		return writeToNBT(new NBTTagCompound());
+	}
+
+	@Override
+	public SPacketUpdateTileEntity getUpdatePacket() {
+		return new SPacketUpdateTileEntity(pos, 1, getUpdateTag());
+	}
+
+	@Override
+	public void onDataPacket(NetworkManager net, SPacketUpdateTileEntity packet) {
+		readFromNBT(packet.getNbtCompound());
+	}
+
+	@Override
+	public Owner getOwner() {
+		return owner;
+	}
+
+	@Override
+	public void setOwner(String uuid, String name) {
+		owner.set(uuid, name);
+	}
+
+	@Override
+	public boolean isModuleEnabled(ModuleType module) {
+		return hasModule(module) && moduleStates.get(module) == Boolean.TRUE; //prevent NPE
+	}
+
+	@Override
+	public void toggleModuleState(ModuleType module, boolean shouldBeEnabled) {
+		moduleStates.put(module, shouldBeEnabled);
+
+		if (shouldBeEnabled)
+			onModuleInserted(getModule(module), module, true);
+		else
+			onModuleRemoved(getModule(module), module, true);
+	}
+}