+++ src/main/java/net/geforcemods/securitycraft/blockentities/ReinforcedDropperBlockEntity.java
@@ -1,0 +1,17 @@
+package net.geforcemods.securitycraft.blockentities;
+
+import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.util.text.ITextComponent;
+
+public class ReinforcedDropperBlockEntity extends ReinforcedDispenserBlockEntity {
+	@Override
+	public String getGuiID() {
+		return "minecraft:dropper";
+	}
+
+	@Override
+	public ITextComponent getDefaultName() {
+		return Utils.localize(SCContent.reinforcedDropper);
+	}
+}